fortivault-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
passlib[bcrypt]
python-jose
pydantic
aiofiles
